fix: Aqua plugin results incorrectly parsed
We're not using the exported ASSURANCE results from the aqua plugin
anymore so we need to use the results file that is generated. This fix
switches to using the results file with the --output flag and cleans up
the removal of the temp files.

Also some house keeping with the makefile