feat: paginated findings
This is a pretty huge but user transparent improvement to the server.
The results from the trivy scan are stored in an LRU with a default of
10 scans being rememebered.

Rather than returning the whole result set (which can be huge) - the
trivy report is broken into findings with a lot of metadata and hints
direct the agent to get the findings from a list endpoint.

Also, added a load of tests - still not much coverage on the MCP related
parts, but this is incredibly difficult to reliably test